convenience target to fetch upstream `octopus/package.py` file
For most merge request to upstream spack, we should start from this file.

So it is good to have a convenient way of getting it, and a convenient way to
check if the current file in this repository is exactly the same as the one in
spack upstream. This targets 'diff' provides that.